Parse nets of pins even with bit-selects.
To that end, collect the necessary information in the Bison rules, when a pin
connection is parsed completely create a Perl-evaluatable string as input for
the pin callback. This string represents the connected nets in form of a hash.

Use this hash later on in dumping, linting, linking and code generation.