Parse nets of pins even with bit-selects
To that end, collect the necessary information in the Bison rules, when a pin
connection is parsed completely create a native Perl hash including msb and lsb
and pass it to a newly created and configurable pinselects callback.
The new parser option is called 'use_bitselects' and defaults to 0 in
Parser and SigParser to avoid regressions.

The code supports some limited parsing of "sized constant numbers" and
parses concatenation elements in pins.

For the code dump the (array of hashes of) nets is serialized in preproc dump.
And linting of nets in case of vectors is slightly diluted:
Warn only if number of usages is higher than number of bits in a vectored net.

Some tests have been modified for coverage and changed output.